+++ src/components/Book/Schedule/addShare/currentList.tsx
@@ -1,15 +1,22 @@
 import React from "react";
 import * as S from "./index.styles";
-import COLORS from "../../../constants/color";
-import getMoneyUnit from "../../../utils/money";
-import { ScheduleProps } from "../../../hooks/book/useAddSchedule";
+import COLORS from "../../../../constants/color";
+import Button from "../../../Button";
+import getMoneyUnit from "../../../../utils/money";
+import { ScheduleProps } from "../../../../hooks/book/useAddSchedule";
 
 export default function CurrentList({
   scheduleList,
   removeSchedules,
+  mutateSchedules,
+  bookId,
+  clearScheduleList,
 }: {
   scheduleList: any;
   removeSchedules: any;
+  mutateSchedules: any;
+  bookId: string | undefined;
+  clearScheduleList: any;
 }) {
   return (
     <S.BankingAddList>
@@ -24,10 +31,17 @@ export default function CurrentList({
                 매달 {schedules.notificationDate}일 /{" "}
                 {schedules.value !== "변동"
                   ? `${getMoneyUnit(Number(schedules.value))}원`
-                  : "변동"}{" "}
+                  : "변동"}
                 /{schedules.isIncome ? <> 수입</> : <> 지출</>}
               </h6>
-              <p>개인 금융 일정</p>
+              <div>
+                {schedules.assignments.map((ass, index) => (
+                  <S.Users key={ass.username}>
+                    {ass.username} {getMoneyUnit(Number(ass.value))}
+                    {index !== schedules.assignments.length - 1 && ` / `}
+                  </S.Users>
+                ))}
+              </div>
             </S.BankingAddInfoWrapper>
             <S.CancelIconButton
               iconSize="2rem"
@@ -40,6 +54,18 @@ export default function CurrentList({
           </S.Box>
         ))}
       </S.BankingAddListWrapper>
+      <Button
+        title="금융 일정 추가하기"
+        theme="quaternary"
+        width="27rem"
+        onClick={() => {
+          mutateSchedules.mutate({
+            id: Number(bookId),
+            schedules: scheduleList,
+          });
+          clearScheduleList();
+        }}
+      />
     </S.BankingAddList>
   );
 }
